Updated models to reflect float values for battery and solar voltages

+++ models/logline.py
@@ -13,8 +13,8 @@ class Logline(Document):
         'pressure': float,
         'humidity': float,
         'light': int,
-        'battery': int,
-        'solar': int
+        'battery': float,
+        'solar': float
     }
     display_fields = ['logtime', 'inT1', 'inT2', 'inT3', 'outT', 'pressure', 'humidity', 'light', 'battery', 'solar']
     field_names = ['Time', "Internal Temp 1 (C)", "Internal Temp 2 (C)", "Internal Temp 3 (C)", "External Temp (C)", "Atmospheric Pressure (mb)", "Humidity (%)", "Ambient Light", "Battery Voltage", "Solar Voltage"]
